Extract out child process stuff
